THIRDEYE-380 Add support to query in-memory tree
Mutable StarTree available via StarTreeManager

Also removed IP_GEO stuff (need to re-think this, write code more cleanly, and
use a better data set)

THIRDEYE-387 selectTreesToQuery bug fix

Before, we were making an assumption that we can use the wall-clock time to
determine which segments should be queried. However, this is incorrect - we can
only use the data time (from segment metadata) to determine which segments to
query, then we can use the wall-clock time (on a per-source basis) to filter
the segments such that we don't double count.

RB=490976
R=kgopalak,npawar
A=kgopalak