[TE] Fix delay when users trigger replay (#1789)
When users trigger replay, the system should respond the job id right away. However, the system wait for long until the detection jobs finish. This is cause by the bug when using Thread. Originally, it uses run(), which triggers the run() in the Runnable instead of triggering in a thread. The correct one should be start(), which generates a thread and triggers the run() in the Runnable.