Fix the package for unit test (#1818)
When running testNG on ThirdEye, this test case isn't executed. It is because of different test package is used in the testBackwardAnomalyFunctionUtils. The test package, org.testng.annotations.Test, is used in all the other unit tests. This pull request is to change the test package in the unit test class.