[TE] Frontend dataset mapping support (#2017)
Currently related metrics and entities are defined on a per-metric basis. While the entity mapping pipeline supports convex hull exploration, this is has side-effects that are hard understand for our users. Instead, they fall back to repetitive entry of similar entity relationships.

This PR introduces a DatasetEntity that logically groups together multiple metrics. Common relationships of a group of metrics can now explicitly be expressed on the dataset level, as well as on an individual per-metric level as before.

* introduced DatasetEntity

* refactored MetricDatasetPipeline

* adaptations to frontend and thirdeye-admin