[TE] cache - replace scheduled reload with on-demand reload (#2730)
This PR replaces eager cache loading with on-demand refresh. The current cache implementation eagerly (re-)loads data for all known data sets. This puts significant load onto the data sources during startup and the (synchronized) refresh of the caches across servers. It is completely unnecessary, however, as all this information could be fetched transparently on-demand instead.