[TE] pinot - support topk limit for groupby queries (#3207)
ThirdEye's Pinot connector currently does not allow us to dynamically change the TOP K limit, which is hard-coded to 100000 elements. This becomes a huge performance burden on de-aggregation queries of data sets with high cardinality, e.g. for rendering the heat map. This PR adds support to limit the number of elements in the response