[TE] rootcause - request chunking for aggregates and scores (#3496)
The RCA dashboard currently loads multiple offsets per metric at once. Unfortunately, this is still insufficient for very large dashboards with 1000+ metrics. This PR adds batching of both metrics and offsets within a single request. Additionally, it enables batching for on-demand scoring of metrics for outliers. This trades off responsiveness for a larger pipe - ultimately this should be superseded by a websocket-like implementation.