[TE] datasource - correction for maxtime offset in pinot datasource (#3598)
This PR fixes the time offset as computed by the pinot data source max time fetcher. Previously, this method computed offsets from epoch without considering the time zone of the underlying data set. Additionally, the PR adds a sanity cutoff for the max time within 1 day in the future from the current timestamp. As a consequence this PR also fixes retrieval of dimension filter keys and values on data sets with a time granularity > 1 week.