[TE] catch exceptions in dimension exploration (#4136)
Currently, the exception in one dimension exploration failed the whole task. The detection should continue to run and save the detected anomalies in other dimensions instead of failing the whole task. Especially when the task is big(exploring thousands of dimensions), the results for other dimensions are wasted and the later catching up task will be more expensive. This PR implements this behavior.