[TE] Evaluation metrics calculation & store (#4202)
The performance metrics need to be stored as a by-product of each detection. Later this will be read by the model evaluators to monitor the detection models. This PR creates the table and adds the necessary boilerplate codes for storing the Evaluation Metrics. The PR also adds the MAPE calculation and collect the metric.