[TE] Fix over-scheduling tasks in data availability scheduler (#4984)
This commit is to fix the issue when detection jobs don't move the watermark forward and the data availability keeps scheduling detection tasks for the same detection. To solve it, we add scheduling window logic in which the scheduler only schedule new detection tasks. If the watermark does not move forward for any reason after the detection job runs, the scheduler will try scheduling it a few times and stop due to being out of the scheduling window.