[TE][subscription] Fix duplicate anomaly report issue and clean up subscription pipelines (#5081)
Bug: Jira alerts were commenting the same anomalies repeatedly as the subscription watermarks were not getting updated in some specific cases.

Cause: This occurs specifically when Jira alerts are subscribed without email. The code had some dependency on email which resulted in the email alerter failing and not updating the subscription watermarks. As a result, the same anomalies were fired again for Jira.

Fix: Made sure the email scheme is not triggered when ONLY jira is configured.

Other Changes:
* Deprecated and removed DIMENSION_ALERTER_PIPELINE in favour of the new DIMENSIONS_ALERTER_PIPELINE which is a super set of the previous.