[TE] add rate limiter to batch alert onboarding (#5234)
Add a rate limiter to the create alert API. The QPS permit can be set in Thirdeye configs. If the QPS is above the permit, it will return an error message to the users. This is to prevent the users who are using API to create alerts at a high throughput which could bring down the system.