[TE] Add validations for the threshold rule anomaly filter (#5278)
The threshold rule anomaly filter is used for filtering the anomaly based on the current value in the anomaly duration.

However, if the metric is not aggregated by SUM or COUNT, the user may set the wrong parameter and cause the system to filter out the anomaly based on the wrong value. This commit fixes the issue.