[TE] Added a backfill start date for Anomaly Detection (#5740)
When an alert is created, it automatically triggers a YAML Onboarding Task
which detects all anomalies in the past 30 days. This change makes the
setting configurable by providing a backfill start timestamp.

Not providing the value or simply providing a dubious entry would result
in the system defaulting to the 30 day lookback.

This change also updates the alert template UI to make it easy for a new
user to try this config

There are no backward incompatible changes.