Simplify compression logic for anomaly timelines view. Now we only keep the values for the selected timestamps after compression, instead of taking an average that could cause value shifts (#5906)
Previously when compressing anomaly timelines view, we first take an average of the group of timestamps to be compressed and put this average to the first timestamp. This naturally would shift the time series leftward, which makes it difficult for recovering the baseline when we want to do anomaly imputation.
Now we only keep the values for the selected timestamps after compression, instead of taking an average that could cause value shifts.