[spi] Breaking Change!! Moving classes in spi to spi package.
All classes inside the spi module now are contained in the spi package. This
is a prerequisite to a plugin class loader implementation.

This is a breaking change on the DB since all the entity names contain the full package along with the class name.

DB Migration:

UPDATE generic_json_entity
SET beanClass = REPLACE(beanClass,
                        'org.apache.pinot.thirdeye.datalayer',
                        'org.apache.pinot.thirdeye.spi.datalayer');