[persistence] Backward Incompatible Refactor. Replaced beanClass with type in generic_json_entity
This is the mother table storing all the data so might be easy to restart dev with a clean slate.
ThirdEye used to store the package + class names in beanClass field which ties the db to the code heavily.
This replaces beanClass with type which is an indirection maintained in the codebase.

Thus, things like beanClass='org.apache.pinot.thirdeye.spi.datalayer.pojo.MergedAnomalyResultBean' is replaced by type='ANOMALY'