[spi] Removing timeConverter interface. no longer used.
History:
Time converter was added to be able to convert to user defined time formats after running the detection. However, once converted to new time formats, these formatted timestamps need to be converted back to epoch time aka longs so that thirdeye can plot them in the UI.
This capability was kept even after the change to handle different use cases. However, now it is dead code and thus removing it.